requirements.txt
colorama==0.4.6
Flask==3.1.1
neovim==0.3.1
python-dotenv==1.1.1
requests==2.32.5
